+++ slack-aws/commands.rb
@@ -2,4 +2,6 @@ require 'slack-aws/commands/about'
 require 'slack-aws/commands/help'
 
 require 'aws-sdk'
+require 'slack-aws/util/aws_client_response'
 require 'slack-aws/commands/s3'
+require 'slack-aws/commands/ops_works'