+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 0.1.0 (Next)
 
-* Added `aws s3 ls [bucket name]` - [@dblock](https://github.com/dblock).
-* Added `aws s3 list buckets` - [@dblock](https://github.com/dblock).
+* Added `aws opsworks instances [stack]` - [@dblock](https://github.com/dblock).
+* Added `aws opsworks apps [stack]` - [@dblock](https://github.com/dblock).
+* Added `aws opsworks stacks` - [@dblock](https://github.com/dblock).
+* Added `aws s3 ls [bucket] [max]` - [@dblock](https://github.com/dblock).
+* Added `aws s3 buckets` - [@dblock](https://github.com/dblock).
 * Initial public release - [@dblock](https://github.com/dblock).